OCR/requirements.txt
flask
pillow
paddlepaddle
paddleocr
opencv-python
numpy
